SD_PRACTICA1/requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1    
redis==5.0.0            
pika==1.3.2    